+++ elastiknn-benchmarks/src/main/scala/com/klibisz/elastiknn/benchmarks/LocalBenchmark.scala
@@ -10,21 +10,38 @@ object LocalBenchmark extends App {
   private val bucket = s"elastiknn-benchmarks"
   private val k = 100
 
-  private val experiments = Seq(
-    Experiment(
-      Dataset.AnnbFashionMnist,
-      Mapping.DenseFloat(Dataset.AnnbFashionMnist.dims),
-      NearestNeighborsQuery.Exact(field, Similarity.L2),
-      Mapping.L2Lsh(Dataset.AnnbFashionMnist.dims, 50, 4, 7),
-      Seq(Query(NearestNeighborsQuery.L2Lsh(field, 1000, 3), k))
-    )
-  )
+  private val experiments =
+    Seq(1).flatMap { shards =>
+      Seq(
+        Experiment(
+          Dataset.AnnbFashionMnist,
+          Mapping.DenseFloat(Dataset.AnnbFashionMnist.dims),
+          Seq(Query(NearestNeighborsQuery.Exact(field, Similarity.L2), k)),
+          shards = shards
+        ),
+        Experiment(
+          Dataset.AnnbFashionMnist,
+          Mapping.L2Lsh(Dataset.AnnbFashionMnist.dims, 75, 4, 7),
+          Seq(
+            Query(NearestNeighborsQuery.L2Lsh(field, 1000 / shards, 0), k),
+            Query(NearestNeighborsQuery.L2Lsh(field, 2000 / shards, 3), k)
+          ),
+          shards = shards
+        ),
+        Experiment(
+          Dataset.AnnbSift,
+          Mapping.L2Lsh(Dataset.AnnbSift.dims, 100, 4, 2),
+          Seq(Query(NearestNeighborsQuery.L2Lsh(field, 5000 / shards, 0), k)),
+          shards = shards
+        )
+      )
+    }
 
   override def run(args: List[String]): URIO[Console, ExitCode] = {
     val s3Url = "http://localhost:9000"
     val s3Client = S3Utils.client(Some(s3Url))
     val experimentEffects = experiments.map { exp =>
-      val key = s"experiments/${exp.md5sum}"
+      val key = s"experiments/${exp.uuid}"
       for {
         _ <- ZIO(s3Client.putObject(bucket, key, codecs.experimentCodec(exp).noSpaces))
         _ <- Execute(
@@ -33,9 +50,7 @@ object LocalBenchmark extends App {
             datasetsPrefix = "data/processed",
             resultsPrefix = "results",
             bucket = bucket,
-            s3Url = Some(s3Url),
-            minWarmupRounds = 5,
-            maxWarmupRounds = 5
+            s3Url = Some(s3Url)
           ))
       } yield ()
     }
@@ -48,9 +63,7 @@ object LocalBenchmark extends App {
           "results",
           "results/aggregate.csv",
           bucket,
-          Some(s3Url),
-          Some("https://api.airtable.com/v0/appmy9gAptPsjo4M7/Results"),
-          sys.env.get("AIRTABLE_API_KEY")
+          Some(s3Url)
         ))
     } yield ()
     pipeline.exitCode